move promptAndTurnOnPublishToAPI
