+++ fluidkeys/keypublish.go
@@ -19,7 +19,10 @@ package main
 
 import (
 	"fmt"
+	"log"
 
+	"github.com/fluidkeys/fluidkeys/colour"
+	"github.com/fluidkeys/fluidkeys/out"
 	"github.com/fluidkeys/fluidkeys/pgpkey"
 )
 
@@ -33,4 +36,17 @@ func publishKeyToAPI(privateKey *pgpkey.PgpKey) error {
 
 	}
 	return nil
+}
+
+func promptAndTurnOnPublishToAPI(prompter promptYesNoInterface, key *pgpkey.PgpKey) {
+	out.Print("🔍 Publishing your key in the Fluidkeys directory allows\n")
+	out.Print("   others to find your key from your email address.\n\n")
+
+	if prompter.promptYesNo(promptPublishToAPI, "", key) == true {
+		if err := Config.SetPublishToAPI(key.Fingerprint(), true); err != nil {
+			log.Printf("Failed to enable publish to api: %v", err)
+		}
+	} else {
+		out.Print(colour.Disabled(" ▸   Not publishing key to API.\n\n"))
+	}
 }
